+++ inlang/source-code/paraglide/paraglide-js/src/cli/defaults.ts
